common/solvers: extract the sparse SPD Cholesky solvers into reusable libs
The sparse Cholesky factor/solve is generic numerical code that does not
belong in an IOP module. Move it out of the highlights harmonic-
transposition code into a new src/common/solvers/ (where choleski.h and
gaussian_elimination.h will follow later):

- common/solvers/sparse_cholesky.h — CPU exact SPD Cholesky (_sp_chol_t +
  _sp_chol_free/_sp_nd_order/_sp_etree/_sp_ereach/_sp_chol_factor/
  _sp_chol_solve). Header-only, like choleski.h; unchanged, so callers
  (_sp_pde_factor, _region_pde_solve, the aniso core) keep working.
- common/solvers/sparse_cholesky_cl.h — GPU level-scheduled SPD Cholesky
  (_sp_chol_cl_t + _sp_chol_cl_free/_sp_cl_upload/_sp_chol_factor_cl/
  _sp_chol_solve_cl). Decoupled from highlights: it no longer takes the
  module's global-data void*; the four data/kernels/highlights_sparse.cl
  kernel handles pass through a small _sp_chol_cl_kernels_t, and the
  _hl_cl_* perf-timing wrappers are replaced with plain dt_opencl_* calls
  (the caller-side factor timing stays). highlights builds the kernel
  struct with a one-line _hl_sp_chol_kernels() helper.

The highlights-specific matrix assembly (_sp_row_*, _sp_pde_*, the
biharmonic dome, the CG fallback) stays in the module and now calls into
the libs. The _sp_chol_t / _sp_chol_cl_t structs move out of
highlights_harmonic_common.h. Dropped the _hl_cl_solve_n diagnostic
counter (its call sites are inside if-conditions; pure perf stat).

Behaviour-neutral: sparse-cl self-test 2.1e-12 (exercises the moved GPU
solver directly), full battery baseline-identical, 6-scene GT identical
CPU+GPU, GCC and clang Debug (-Werror) green.


The new solver headers (sparse_cholesky.h / sparse_cholesky_cl.h) are
clang-formatted here too (folded in from the former standalone format
commit).
